+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHQZ3W9KP8KDAT4CDZM1D9P4"
+        "01KHRCGAHAHJ861EZ02YYC0ZB6"
     )
     r"""The unique identifier of the data source"""