+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHQZ3W9M716H9AP41YX387ZA"
+        "01KHRCGAHCPCQ0E67CB1P5H2TC"
     )
     r"""The unique identifier of the data source"""