+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHQZ3W9Q2013JMZH3TW6A4JD"
+        "01KHRCGAHR6G6K2321GKX7ADDR"
     )
     r"""The unique identifier of the data source"""