+++ packages/orq-rc/RELEASES.md
@@ -5779,3 +5779,13 @@ Based on:
 - [python v4.4.0-rc.9] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.9] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.9 - packages/orq-rc
+
+## 2026-02-18 12:48:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.722.2 (2.832.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.10] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.10] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.10 - packages/orq-rc